On branch master Your branch is up to date with 'origin/master'.
Changes to be committed:
	new file:   all.midd
	modified:   hc.cpp
	new file:   ip.midd
	new file:   ipttl.midd

+++ hc.cpp
@@ -6,7 +6,7 @@ HcFilter::HcFilter(std::string filename)
     std::string ip;
     v6addr v6;
     uint32_t v4;
-    uint8_t ttl;
+    int ttl;
     while(in.good()){
         in >> ip >> ttl;
         if(ip.find(':') != std::string::npos){
